+++ apps/calendar/src/modules/template.ts
@@ -1,9 +1,7 @@
 import { registerTemplateConfig } from '@src/template';
 import { InitStoreData } from '@t/store';
 
-const template = {
+export const template = {
   name: 'template',
   state: ({ options }: InitStoreData) => registerTemplateConfig(options?.template),
 };
-
-export default template;
