+++ apps/calendar/src/modules/theme.ts
@@ -1,9 +1,7 @@
 import Theme from '@src/theme';
 import { InitStoreData } from '@t/store';
 
-const theme = {
+export const theme = {
   name: 'theme',
   state: ({ options }: InitStoreData) => new Theme(options.theme),
 };
-
-export default theme;
